+++ tests-polyfill/w3c/IDBObjectStore.add.js
@@ -464,7 +464,7 @@ describe('W3C IDBObjectStore.add Tests', function () {
     });
 
     // value_recursive
-    it('Recursive value', function (done) {
+    it.skip('Recursive value', function (done) {
         var count = 0;
 
         function recursive_value(desc, value) {